Put testing generator into test mod, mixin rename
