+++ src/test/resources/sltest.mixins.json
@@ -4,6 +4,7 @@
   "package": "net.modificationstation.sltest.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
+    "MixinOverworldLevelSource"
   ],
   "server": [
   ],
