+++ station-flattening-v0/src/main/resources/station-flattening-v0.mixins.json
@@ -15,7 +15,7 @@
     "MixinLevelProperties",
     "MixinWorldPopulationRegion",
     "MixinBlock",
-    "MixinTestOverworldLevelSource"
+	"MixinOverworldLevelSource"
   ],
   "client": [
     "MixinClientLevel",
